refactor: rename `createConfig` to `defineConfig`
