+++ dev/design-studio/sanity.config.ts
@@ -1,11 +1,11 @@
-import {createConfig} from 'sanity'
+import {defineConfig} from 'sanity'
 import {deskTool} from 'sanity/desk'
 import {templates} from './templates'
 import {themePreviewTool} from './plugins/theme-preview-tool'
 import {schemaTypes} from './schemaTypes'
 import {structure} from './structure'
 
-export default createConfig({
+export default defineConfig({
   plugins: [
     deskTool({structure}),
     themePreviewTool(),
