+++ dev/starter-studio/sanity.cli.ts
@@ -1,6 +1,6 @@
-import {createCliConfig} from 'sanity/cli'
+import {defineCliConfig} from 'sanity/cli'
 
-export default createCliConfig({
+export default defineCliConfig({
   api: {
     projectId: 'ppsg7ml5',
     dataset: 'test',
