+++ examples/blog-studio/sanity.config.ts
@@ -1,8 +1,8 @@
-import {createConfig} from 'sanity'
+import {defineConfig} from 'sanity'
 import {deskTool} from 'sanity/desk'
 import {schemaTypes} from '../../packages/@sanity/cli/templates/blog/schemas'
 
-export default createConfig({
+export default defineConfig({
   name: 'default',
   title: 'Blog studio example',
