+++ packages/@sanity/cli/src/actions/init-project/createCliConfig.ts
@@ -3,9 +3,9 @@ import * as parser from 'recast/parsers/typescript'
 import traverse from '@babel/traverse'
 
 const defaultTemplate = `
-import {createCliConfig} from 'sanity/cli'
+import {defineCliConfig} from 'sanity/cli'
 
-export default createCliConfig({
+export default defineCliConfig({
   api: {
     projectId: '%projectId%',
     dataset: '%dataset%'
