+++ packages/@sanity/cli/src/actions/init-project/templates/ecommerce.ts
@@ -1,12 +1,12 @@
 import type {ProjectTemplate} from '../initProject'
 
 const configTemplate = `
-import {createConfig} from 'sanity'
+import {defineConfig} from 'sanity'
 import {deskTool} from 'sanity/desk'
 import {barcodeInput} from './plugins/barcode
 import {schemaTypes} from './schemas'
 
-export default createConfig({
+export default defineConfig({
   name: '%sourceName%',
   title: '%projectName%',
