+++ packages/@sanity/cli/src/actions/init-project/templates/shopify.ts
@@ -4,14 +4,14 @@ import type {ProjectTemplate} from '../initProject'
 // @todo new document structure
 // @todo document actions
 const configTemplate = `
-import {createConfig} from 'sanity'
+import {defineConfig} from 'sanity'
 import {deskTool} from 'sanity/desk'
 import {dashboard} from '@sanity/dashboard'
 import {media} from 'sanity-plugin-media'
 import {structure} from './deskStructure'
 import {schemaTypes} from './schemas'
 
-export default createConfig({
+export default defineConfig({
   plugins: [
     deskTool({structure}),
     media(),
