+++ packages/@sanity/cli/test/__fixtures__/v3/sanity.cli.ts
@@ -1,6 +1,6 @@
-import {createCliConfig} from 'sanity/cli'
+import {defineCliConfig} from 'sanity/cli'
 
-export default createCliConfig({
+export default defineCliConfig({
   api: {
     projectId: 'aeysrmym',
     dataset: 'production',
