+++ packages/@sanity/cli/test/__fixtures__/v3/sanity.config.ts
@@ -1,9 +1,9 @@
-import {createConfig} from 'sanity'
+import {defineConfig} from 'sanity'
 import {deskTool} from 'sanity/desk'
 import {MyLogo} from './components/MyLogo'
 import {schema} from './schema'
 
-export default createConfig({
+export default defineConfig({
   name: 'default', // @todo remove
   projectId: 'aeysrmym',
   dataset: 'production',
