+++ packages/sanity/exports/cli.ts
@@ -1,3 +1,3 @@
-export {createCliConfig, getCliClient} from '@sanity/cli'
+export {createCliConfig, defineCliConfig, getCliClient} from '@sanity/cli'
 
 export type {CliClientOptions, CliConfig} from '@sanity/cli'
