--- packages/sanity/src/core/config/createConfig.ts
@@ -1,6 +1,0 @@
-import type {Config} from './types'
-
-/** @beta */
-export function createConfig<T extends Config>(config: T): T {
-  return config
-}
